requirements.txt
docker-py
docker-pycreds
docutils
kafka-python
neo4j-driver
numpy
pandas
pymongo
Pypubsub
python-dateutil
requests
scikit-learn
scipy
sklearn
xmltodict
tika
nltk
networkx
